requirements.txt
asgiref==3.11.0
Django==5.2.8
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.4
requests==2.31.0
